Added BitcoinTalk generator
This generator create a raw data file of around 900 MB; its processing requires more than 4 GB RAM. The download of the whole data, while remaining compliant to the access policy of the BitcoinTalk forum, which requires a pause of 1 second between two subsequent page downloads, takes currently more than 40 days.

The generator enables to interrupt download of data; if started with the command line argument `update`, it calculates the maximal ID number and start downloading data with the next ID.

As the generator counts up the ID numbers, it ends after downloading 1000 subsequent empty (i.e. "user not found") pages. Then it calculates, which pages are missing, and tries to download them; this way the generator tries to compensate forcibly disconnected Internet connexions. Currently, there 1942 empty pages, so it takes a little more than 30 minutes to get through them.

Some improvements were made to the BCH regex (a \b was inserted before the second part of it), else it produces many false positives both within BTC and with ETH addresses.

The data download part was made more robust, as it retries downloads if encountering timeouts.