Made data collection parallel
This implements issue #20. Data from all three source is collected simultaneously. Data from transactions log is overwritten, as both leaderboards (two other sources) are connected to certain dates and thus do not change with time.

Further, it seems that the web servers of PipeFlare.io are hidden behind CloudFlare. This presents a problem, because the referral leaderboard pages are calculated slowly, and hence CloudFlare returns the page for error code 524 (timeout). This was fixed through a fixed maximal number of retries, which is currently 20.

Also, the input ZEC address collection prints warning, if collected TX hash does not exist (yet), and proceeds with the next TX link.